+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# d0003 - 2024-01-10
+* Search service:
+  * Added support of `motodart.ru` provider
+
 # d0002 - 2024-01-02
 * Reworked search service endpoints
   * Moved GUI from `/api/search` to `/search` endpoint